Changed file names and added transference num
